R&D/requirements.txt
langchain-core
langchain-community
langchain
PyPDF2
python-docx
pandas
beautifulsoup4
lxml
requests
pymongo
langchain-mongodb
langchain-huggingface